+++ mrmustard/lab_dev/circuits.py
@@ -21,8 +21,11 @@ A class to quantum circuits.
 from __future__ import annotations
 
 from collections import defaultdict
+from pydoc import locate
 from typing import Optional, Sequence, Union
+
 from mrmustard import math, settings
+from mrmustard.utils.serialize import save
 from mrmustard.lab_dev.circuit_components import CircuitComponent
 from mrmustard.lab_dev.transformations import BSgate
 
@@ -426,6 +429,36 @@ class Circuit:
                     raise ValueError(f"The contraction ``{i1, i2}`` is invalid.")
 
             remaining[i1] = (remaining[i1] @ remaining.pop(i2))[0]
+
+    def serialize(self, filestem: str = None):
+        r"""
+        Serialize a Circuit.
+
+        Args:
+            filestem: An optional name to give the resulting file saved to disk.
+        """
+        components, data = list(zip(*[c._serialize() for c in self.components]))
+        kwargs = {
+            "arrays": {f"{k}:{i}": v for i, arrs in enumerate(data) for k, v in arrs.items()},
+            "path": self.path,
+            "components": components,
+        }
+        return save(type(self), filename=filestem, **kwargs)
+
+    @classmethod
+    def deserialize(cls, data: dict) -> Circuit:
+        r"""Deserialize a Circuit."""
+        comps, path = map(data.pop, ("components", "path"))
+
+        for k, v in data.items():
+            kwarg, i = k.split(":")
+            comps[int(i)][kwarg] = v
+
+        classes: list[CircuitComponent] = [locate(c.pop("class")) for c in comps]
+        circ = cls([c._deserialize(comp_data) for c, comp_data in zip(classes, comps)])
+        if path:  # re-evaluates the hidden `_graph` property
+            circ.path = [tuple(p) for p in path]
+        return circ
 
     def __eq__(self, other: Circuit) -> bool:
         return self.components == other.components